+++ packages/steiner-demo/src/modules/hotels/actions/hotels.js
@@ -1,5 +1,3 @@
-// import { actionCreator } from 'steiner';
-
 import helper from 'helpers/steinerHelper';
 
 export const actionTypes = helper.createActionTypes('hotels');