+++ packages/steiner/src/helpers/actionCreator.js
@@ -1,11 +1,13 @@
 import arrify from 'arrify';
 import upperFirst from 'lodash/upperFirst';
-import defaults from 'lodash/defaults';
-import forOwn from 'lodash/forOwn';
+import get from 'lodash/get';
+// import defaults from 'lodash/defaults';
+// import forOwn from 'lodash/forOwn';
 import trim from 'lodash/trim';
 import template from 'lodash/template';
 
-import defaultMessages from '../messages/en';
+// import defaultMessages from '../messages/en';
+import { getTranslations, getLanguage } from '../settings/reducer';
 
 function addAsyncGroup(resource, actionTypes, group, options) {
     const upperGroup    = group.toUpperCase();
@@ -105,48 +107,82 @@ export function createActionTypes(resource, options = {}) {
     return actionTypes;
 }
 
-function generateSuccessNotification(message, title) {
-    return {
-        title,
-        message,
-        status: 'success'
+// function generateSuccessNotification(message, title) {
+//     return {
+//         title,
+//         message,
+//         status: 'success'
+//     }
+// }
+
+// function generateFailNotification(message, title) {
+//     return {
+//         title,
+//         message,
+//         status: 'error'
+//     };
+// }
+
+// export function createDefaultMessages(resource, messageTemplates = {}) {
+//     const messages = {};
+
+//     defaults(messageTemplates, defaultMessages.templates.actionMessages);
+
+//     forOwn(messageTemplates, (value, key) => {
+//         if (typeof value === 'function') {
+//             messages[key] = value({ resource });
+//         } else if (typeof value === 'string') {
+//             const compiled = template(value, {
+//                 interpolate : /\{\{([\s\S]+?)\}\}/g
+//             });
+//             messages[key] = compiled({ resource });
+//         }
+//     });
+
+//     return messages;
+// }
+
+// export function createActionMessages(resource, options = {}) {
+//     const messages = options.messages || {};
+
+//     const defaultMessages = createDefaultMessages(resource, options.messageTemplates);
+//     defaults(messages, defaultMessages);
+
+//     return messages;
+// }
+
+function generateNotificationPayload(actionKey, type, messages, titles, getState, resource) {
+    let translations;
+
+    const state = getState();
+    const language = getLanguage(state);
+
+    let message = get(messages, `${language}.${actionKey}`);
+    let title = get(titles, `${language}.${type}`);
+
+    if (! message) {
+        translations = getTranslations(state);
+
+        const compiled = template(translations.templates.actionMessages[actionKey], {
+            interpolate : /\{\{([\s\S]+?)\}\}/g
+        });
+
+        message = compiled({ resource })
     }
-}
 
-function generateFailNotification(message, title) {
+    if (! title) {
+        if (! translations) {
+            translations = getTranslations(state);
+        }
+
+        title = translations.messages.notifications.titles.success;
+    }
+
     return {
         title,
         message,
-        status: 'error'
-    };
-}
-
-export function createDefaultMessages(resource, messageTemplates = {}) {
-    const messages = {};
-
-    defaults(messageTemplates, defaultMessages.templates.actionMessages);
-
-    forOwn(messageTemplates, (value, key) => {
-        if (typeof value === 'function') {
-            messages[key] = value({ resource });
-        } else if (typeof value === 'string') {
-            const compiled = template(value, {
-                interpolate : /\{\{([\s\S]+?)\}\}/g
-            });
-            messages[key] = compiled({ resource });
-        }
-    });
-
-    return messages;
-}
-
-export function createActionMessages(resource, options = {}) {
-    const messages = options.messages || {};
-
-    const defaultMessages = createDefaultMessages(resource, options.messageTemplates);
-    defaults(messages, defaultMessages);
-
-    return messages;
+        status: type === 'success' ? 'success' : 'error'
+    }
 }
 
 export function createActions(resource, actionTypes, options = {}) {
@@ -156,8 +192,10 @@ export function createActions(resource, actionTypes, options = {}) {
 
     resource = trim(upperFirst(resource));
 
-    const messages = createActionMessages(resource, options);
-    const titles = options.notificationTitles || defaultMessages.messages.notifications.titles;
+    // const messages = createActionMessages(resource, options);
+    // const titles = options.notificationTitles || defaultMessages.messages.notifications.titles;
+    const messages = options.messages || {};
+    const titles = options.notificationTitles || {};
 
     const actions = {};
 
@@ -187,19 +225,23 @@ export function createActions(resource, actionTypes, options = {}) {
     }
 
     actions['createSuccess'] = function(response) {
-        return {
-            type: actionTypes.createSuccess,
-            payload: response,
-            notification: generateSuccessNotification(messages.createSuccess, titles.success)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.createSuccess,
+                payload: response,
+                notification: generateNotificationPayload('createSuccess', 'success', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions['createFail'] = function(error) {
-        return {
-            type: actionTypes.createFail,
-            error,
-            notification: generateFailNotification(messages.createFail, titles.fail)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.createFail,
+                error,
+                notification: generateNotificationPayload('createFail', 'fail', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions[`update`] = function(id, data) {
@@ -213,19 +255,23 @@ export function createActions(resource, actionTypes, options = {}) {
     }
 
     actions['updateSuccess'] = function(response) {
-        return {
-            type: actionTypes.updateSuccess,
-            payload: response,
-            notification: generateSuccessNotification(messages.updateSuccess, titles.success)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.updateSuccess,
+                payload: response,
+                notification: generateNotificationPayload('updateSuccess', 'success', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions['updateFail'] = function(error) {
-        return {
-            type: actionTypes.updateFail,
-            error,
-            notification: generateFailNotification(messages.updateFail, titles.fail)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.updateFail,
+                error,
+                notification: generateNotificationPayload('updateFail', 'fail', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions[`delete`] = function(id) {
@@ -238,19 +284,23 @@ export function createActions(resource, actionTypes, options = {}) {
     }
 
     actions['deleteSuccess'] = function(payload) {
-        return {
-            type: actionTypes.deleteSuccess,
-            payload,
-            notification: generateSuccessNotification(messages.deleteSuccess, titles.success)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.deleteSuccess,
+                payload,
+                notification: generateNotificationPayload('deleteSuccess', 'success', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions['deleteFail'] = function(error) {
-        return {
-            type: actionTypes.deleteFail,
-            error,
-            notification: generateFailNotification(messages.deleteFail, titles.fail)
-        };
+        return (dispatch, getState) => {
+            dispatch({
+                type: actionTypes.deleteFail,
+                error,
+                notification: generateNotificationPayload('deleteFail', 'fail', messages, titles, getState, resource)
+            });
+        }
     }
 
     actions[`resetCurrent`] = function() {